+++ package.json
@@ -19,10 +19,13 @@
   },
   "dependencies": {
     "@element-plus/icons-vue": "^2.1.0",
+    "@types/marked": "^5.0.1",
     "axios": "^1.3.5",
     "echarts": "^5.4.2",
     "echarts-liquidfill": "^3.1.0",
     "element-plus": "^2.3.3",
+    "highlight.js": "^11.8.0",
+    "highlightjs-line-numbers.js": "^2.8.0",
     "lodash": "^4.17.21",
     "marked": "^5.1.1",
     "moment": "^2.29.4",
