+++ src/api/index.ts
@@ -1,5 +1,5 @@
 import * as blog from './blog'
-import * as login from '@/api/login'
+import * as login from './login'
 export default {
   blog,
   login,
