+++ src/main.ts
@@ -9,6 +9,7 @@ import ElementPlus from 'element-plus'
 import 'element-plus/dist/index.css'
 //@ts-expect-error忽略当前文件ts类型的检测否则有红色提示(打包会失败)
 import zhCn from 'element-plus/dist/locale/zh-cn.mjs'
+import 'highlight.js/styles/foundation.css'
 
 const app = createApp(App)
 app.use(ElementPlus, {
