+++ src/router/routers.ts
@@ -1,5 +1,5 @@
-import { createRouter, createWebHashHistory } from "vue-router";
-import routes from "./routes";
+import { createRouter, createWebHashHistory } from 'vue-router'
+import routes from './routes'
 
 const router = createRouter({
   history: createWebHashHistory(),
@@ -9,23 +9,23 @@ const router = createRouter({
     return {
       left: 0,
       top: 0,
-    };
+    }
   },
-});
+})
 
 // 前置守卫
-router.beforeEach((to, from, next) => {
-  const token = localStorage.getItem("token");
-  if (!token && to.path !== "/login") {
-    next("/login");
-  } else {
-    next();
-  }
-});
+// router.beforeEach((to, from, next) => {
+//   const token = localStorage.getItem('token')
+//   if (!token && to.path !== '/login') {
+//     next('/login')
+//   } else {
+//     next()
+//   }
+// })
 
 // // 后置守卫
 // router.afterEach((to, from) => {
 //   // console.log(to, from)
 // })
 
-export default router;
+export default router
