+++ src/router/routes.ts
@@ -1,19 +1,24 @@
-import { RouteRecordRaw } from "vue-router";
+import { RouteRecordRaw } from 'vue-router'
 
+// 前台页面
 const welcome = () => import('@/views/welcome.vue')
 const login = () => import('@/views/login.vue')
 const home = () => import('@/views/home.vue')
 const Essay = () => import('@/views/Essay.vue')
+const detail = () => import('@/views/detail.vue')
+
+// 后台管理页面
 const admin = () => import('@/views/admin/index.vue')
 // 路由配置
 const routes: RouteRecordRaw[] = [
   {
-    path: "/",
+    path: '/',
     component: welcome,
-    redirect: "/home",
+    redirect: '/home',
     children: [
       { path: '/home', component: home },
       { path: '/essay', component: Essay },
+      { path: '/detail', component: detail },
     ],
   },
   {
@@ -26,4 +31,4 @@ const routes: RouteRecordRaw[] = [
   },
 ]
 
-export default routes;
+export default routes
