+++ src/views/detail.vue
@@ -1,0 +1,37 @@
+<template>
+  <el-input type="textarea" v-model="value" @input="change"></el-input>
+  <div v-html="markdownToHtml" class="markdown-body"></div>
+</template>
+
+<script setup lang="ts">
+import { marked } from 'marked'
+import { ref, shallowRef, onMounted } from 'vue'
+import hljs from 'highlight.js'
+import 'highlight.js/styles/monokai-sublime.css'
+import api from '@/api'
+
+onMounted(() => {
+  api.blog.getBlogDetail().then((res) => {
+    console.log(res)
+  })
+})
+
+const render = new marked.Renderer()
+marked.setOptions({
+  renderer: render, // 这是必填项
+  gfm: true, // 启动类似于Github样式的Markdown语法
+  pedantic: false, // 只解析符合Markdwon定义的，不修正Markdown的错误
+  sanitize: false, // 原始输出，忽略HTML标签（关闭后，可直接渲染HTML标签）
+  // 高亮的语法规范
+  highlight: (code, lang) => hljs.highlight(code, { language: lang }).value,
+})
+const value = ref('')
+const markdownToHtml = shallowRef('')
+markdownToHtml.value = marked(value.value)
+
+const change = (value: string) => {
+  markdownToHtml.value = marked(value)
+}
+</script>
+
+<style lang="scss" scoped></style>
